fix other retention values

+++ src/LEGO.AsyncAPI.Bindings/Kafka/KafkaChannelBinding.cs
@@ -48,8 +48,8 @@ namespace LEGO.AsyncAPI.Bindings.Kafka
         {
             { "cleanup.policy", (a, n) => { a.CleanupPolicy = n.CreateSimpleList(s => s.GetScalarValue()); } },
             { "retention.ms", (a, n) => { a.RetentionMilliseconds = n.GetLongValue(); } },
-            { "retention.bytes", (a, n) => { a.RetentionBytes = n.GetIntegerValue(); } },
-            { "delete.retention.ms", (a, n) => { a.DeleteRetentionMilliseconds = n.GetIntegerValue(); } },
+            { "retention.bytes", (a, n) => { a.RetentionBytes = n.GetLongValue(); } },
+            { "delete.retention.ms", (a, n) => { a.DeleteRetentionMilliseconds = n.GetLongValue(); } },
             { "max.message.bytes", (a, n) => { a.MaxMessageBytes = n.GetIntegerValue(); } },
             { "confluent.key.schema.validation", (a, n) => { a.ConfluentKeySchemaValidation = n.GetBooleanValue(); } },
             { "confluent.key.subject.name.strategy", (a, n) => { a.ConfluentKeySubjectName = n.GetScalarValue(); } },
